+++ lib/larva/worker_pool.rb
@@ -11,7 +11,7 @@ module Larva
 
     def start
       start_workers
-      keep_workers_alive if @workers.count > 0
+      keep_workers_alive if workers.count > 0
     end
 
     private
@@ -22,7 +22,7 @@ module Larva
         worker[:name] = "Listener for #{topic}"
         worker
       end
-      logger.info "#{@workers.count} threads started."
+      logger.info "#{workers.count} threads started."
     end
 
     def start_worker(topic, processor)
@@ -32,9 +32,9 @@ module Larva
     end
 
     def keep_workers_alive
-      sleep(5) while @workers.all? { |t| t.alive?  }
+      sleep(5) while workers.all? { |t| t.alive?  }
       logger.error 'Some threads have died:'
-      @workers.each do |worker|
+      workers.each do |worker|
         logger.error "#{worker[:name]} was #{worker.alive? ? 'alive' : 'dead'}"
       end
     end